backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
tensorflow==2.6.0
torch==1.9.0
rasterio==1.2.10
geopandas==0.9.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
requests==2.26.0
aiohttp==3.8.1
python-multipart==0.0.5
sqlalchemy==1.4.23
alembic==1.7.1
pytest==6.2.5
pytest-asyncio==0.15.1
httpx==0.19.0 